Fix memory corruption bug where wrong sized value was being used to store pointers (breaks on 64-bit machines with randomised address spaces).

+++ glsl-frontend/glslang/MachineIndependent/preprocessor/memory.c
@@ -79,15 +79,23 @@ NVIDIA HAS BEEN ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <stdlib.h>
 #include <string.h>
 
+// RJW:
+// The definition of uintptr_t was buggy in the original code.
+// Using stdint.h makes us less portable but more likely to be
+// correct.
+#include <stdint.h>
+
+/*
 #ifdef __STDC99__
 #include <stdint.h>
 #elif defined (_WIN64)
 typedef unsigned __int64 uintptr_t;
 #elif defined (linux)
-typedef unsigned int uintptr_t;
+typedef uint64_t uintptr_t;
 #else
 // uintptr_t defined on MacOS.
 #endif
+*/
 
 #include "memory.h"
